Extended toplevel. Fix not needing commas sometimes
It accepted some data where a comma was not present after a first
field in a record and tuple. That's fixed.

Also here comes the big difference from ron-rs: you can now use record
and list at toplevel without brackets, like this:

    field1: "hello",
    field2: "world!"

or like this:

    "hello",
    Val ("list", "at"),
    Toplevel()

Those are valid ron-hs documents that you can parse! And the values
don't have to be on new lines, it's still not whitespace significant