Fix not parsing floats according to spec
But actually there's a bigger problem that ron-rs doesn't parse its
floats according to spec. They can parse:
- NaN (which should be an identifier)
- inf (ditto)
- -inf (which is illegal)
- 1e1 - illegal because a float should have a . in its representation. I
  think I want to parse this in the future, it's a convenient notation

Now ron-rs can also parse negative zero. We can parse it, but as a
positive zero because of Data.Scientific, so it's unclear which way is
better

What's fun about inf and NaN is that ron-rs can /also/ parse them as
identifiers, depending on what the deserialization requires. I wonder
how it works in ron-rewrite, as they use a separate tree phase.